+++ product-service/src/main/java/com/rent/project/productservice/repository/ProductDetailRepo.java
@@ -4,8 +4,8 @@ import com.rent.project.productservice.models.ProductDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ProductDetailRepo extends JpaRepository<ProductDetail,Integer> {
+public interface ProductDetailRepo extends JpaRepository<ProductDetail,Long> {
 
     @Query(value = "select * from product_detail where product_id=?1",nativeQuery = true)
-    ProductDetail fetchProductDetail(Integer id);
+    ProductDetail fetchProductDetail(Long id);
 }
