+++ product-service/src/main/java/com/rent/project/productservice/repository/ProductDurationRateRepo.java
@@ -3,5 +3,5 @@ package com.rent.project.productservice.repository;
 import com.rent.project.productservice.models.ProductDurationRate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductDurationRateRepo extends JpaRepository<ProductDurationRate,Integer> {
+public interface ProductDurationRateRepo extends JpaRepository<ProductDurationRate,Long> {
 }
