+++ product-service/src/main/java/com/rent/project/productservice/repository/UserCredentialsRepo.java
@@ -3,5 +3,5 @@ package com.rent.project.productservice.repository;
 import com.rent.project.productservice.models.UserCredentials;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserCredentialsRepo extends JpaRepository<UserCredentials,Integer> {
+public interface UserCredentialsRepo extends JpaRepository<UserCredentials,Long> {
 }
